chore: remove duplicated hooks
